requirements.txt
boto3~=1.5.3
pyTenable~=0.3.0
tenable-io~=1.4.1
